Removes xd dsp; updates scss variables and globals
Adobe XD DSP was a bit too much overhead for me at this point, so I just removed it.

New styling for the site has been captured in new variables and applied globally to the site for the elements that make sense to do that for: headings, ul/li, etc. Will break out global styling in future refactoring if needed.